reviewed

+++ Firebird.2.5/Tests/Linq/ConvertTests/Tests.Linq.ConvertTests.ConvertFromOneToAnother(Firebird.2.5).sql
@@ -30,7 +30,7 @@ DECLARE @value Decimal(9, 5)
 SET     @value = 6579.64648
 
 SELECT FIRST 1
-	CAST(@value AS Float)
+	CAST(@value AS DOUBLE PRECISION)
 FROM
 	"LinqDataTypes" "t1"
 
@@ -39,7 +39,7 @@ DECLARE @value Float -- Single
 SET     @value = 6579.64648
 
 SELECT FIRST 1
-	CAST(@value AS Float)
+	CAST(@value AS DOUBLE PRECISION)
 FROM
 	"LinqDataTypes" "t1"
 
@@ -48,7 +48,7 @@ DECLARE @value Double
 SET     @value = 6579.6464800000003
 
 SELECT FIRST 1
-	CAST(@value AS Float)
+	CAST(@value AS DOUBLE PRECISION)
 FROM
 	"LinqDataTypes" "t1"
