+++ Firebird.2.5/Tests/Linq/ConvertTests/Tests.Linq.ConvertTests.ConvertToDouble(Firebird.2.5).sql
@@ -5,7 +5,7 @@ SELECT
 FROM
 	(
 		SELECT
-			CAST("t"."MoneyValue" AS Float) as "c1"
+			CAST("t"."MoneyValue" AS DOUBLE PRECISION) as "c1"
 		FROM
 			"LinqDataTypes" "t"
 	) "p"
