+++ Firebird.3/Tests/Linq/ConditionalTests/Tests.Linq.ConditionalTests.Conditional_DoNotEvalueBothBranches(Firebird.3).sql
@@ -1,6 +1,6 @@
 ﻿-- Firebird.3 Firebird3
 
 SELECT
-	CAST(20 AS Float)
+	CAST(20 AS DOUBLE PRECISION)
 FROM rdb$database
