+++ Firebird.3/Tests/Linq/MathFunctionTests/Tests.Linq.MathFunctionTests.Truncate2(Firebird.3).sql
@@ -5,7 +5,7 @@ SELECT
 FROM
 	(
 		SELECT
-			Trunc(CAST(-"p"."MoneyValue" AS Float), 0) as "Value_1"
+			Trunc(CAST(-"p"."MoneyValue" AS DOUBLE PRECISION), 0) as "Value_1"
 		FROM
 			"LinqDataTypes" "p"
 	) "t"
