+++ Firebird.4/Tests/Linq/ConvertTests/Tests.Linq.ConvertTests.ToDouble(Firebird.4).sql
@@ -5,5 +5,5 @@ SELECT
 FROM
 	"LinqDataTypes" "p"
 WHERE
-	CAST(Floor(CAST("p"."MoneyValue" AS Float)) AS Int) > 0
+	CAST(Floor(CAST("p"."MoneyValue" AS DOUBLE PRECISION)) AS Int) > 0
