+++ Firebird.5/Tests/Linq/MappingTests/Tests.Linq.MappingTests.MappingTypingByConstant_FromEnumerable_Double(Firebird.5,0.0d).sql
@@ -6,7 +6,7 @@ SELECT
 FROM
 	"Person" "entity"
 		INNER JOIN (
-			SELECT 1 AS "Id", CAST(0 AS Float) AS "Value" FROM rdb$database) "arg" ON "entity"."PersonID" = "arg"."Id"
+			SELECT 1 AS "Id", CAST(0 AS DOUBLE PRECISION) AS "Value" FROM rdb$database) "arg" ON "entity"."PersonID" = "arg"."Id"
 
 -- Firebird.5 Firebird4
 
@@ -16,5 +16,5 @@ SELECT
 FROM
 	"Person" "entity"
 		INNER JOIN (
-			SELECT 1 AS "Id", CAST(3147483648 AS Float) AS "Value" FROM rdb$database) "arg" ON "entity"."PersonID" = "arg"."Id"
+			SELECT 1 AS "Id", CAST(3147483648 AS DOUBLE PRECISION) AS "Value" FROM rdb$database) "arg" ON "entity"."PersonID" = "arg"."Id"
